requirements.txt
asgiref==3.8.1
blinker==1.8.2
build==1.2.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dj-database-url==2.2.0
Django==4.2
django-cron==0.6.0
django-tailwind==3.8.0
fire==0.6.0
gunicorn==22.0.0
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
packaging==24.1
pillow==10.4.0
pip-tools==7.4.1
psycopg2==2.9.9
Pygments==2.18.0
pyproject_hooks==1.1.0
python-decouple==3.8
python-dotenv==1.0.1
PyYAML==6.0.1
rav==0.0.9
requests==2.32.3
rich==13.7.1
six==1.16.0
sqlparse==0.5.0
stripe==10.4.0
termcolor==2.4.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.5
Werkzeug==3.0.3
whitenoise==6.7.0
schedule
